Editing out sRGB assumption in color management basic operations
Alternative raw channel scaling and highlight clipping when using white balance

Changing the type of tiff_float_size from int to double to avoid unexpected rounding during ICC profile loading

Fixing some aftermath after r5.6-dev rebase

Adding conditional control for image processing withut silen assumption of sRGB profile through application settings flag - 'Don't silently assume sRGB for ICC processing'

+++ rtengine/dcp.cc
@@ -476,7 +476,7 @@ DCPProfile::DCPProfile(const Glib::ustring& filename) :
     baseline_exposure_offset(0.0)
 {
     delta_info.hue_step = delta_info.val_step = look_info.hue_step = look_info.val_step = 0;
-    constexpr int tiff_float_size = 4;
+    constexpr double tiff_float_size = 4.0f;
 
     enum class TagKey : int {
         COLOR_MATRIX_1 = 50721,
